single shot reduction wrapper, in progress

+++ single_shot_reduction/reduce_shot.py
@@ -27,12 +27,18 @@ from dataclasses import dataclass
 import traceback
 
 
+DEBUG_NO_COPY = False
+
 ########################################################################
 # CONFIGURATION
 ########################################################################
 FutureShotDateLimit = 20490101000  # do not allow shots after this dave+shot
 
 ScriptRepo = "/work/03261/polonius/hetdex/single_shot"
+LocalScriptRepo = "./local_script_repo" #useful if running multiple single shots ... can copy remotely once
+                                        #then copy locally from here for each shot
+                                        #set to None if you do NOT want to use a local script dir cache
+                                        #  and force a copy from the main repo each time
 
 WorkDirRoot = "./"
 
@@ -52,6 +58,7 @@ class Config:
     datevshot: str
     cwd_orig: str
     cwd: str
+    scriptdir: str
 
 
 ########################################################################
@@ -62,7 +69,7 @@ args = list(sys.argv) #python3 map is no longer a list, so need to cast here
 args.pop(0) #remove THIS file
 args = [x.replace("--","-") for x in args]
 
-cfg = Config(False,False,0,"",os.getcwd(),os.getcwd())
+cfg = Config(False,False,0,"",os.getcwd(),os.getcwd(),"")
 
 if "-clean" in args:
     cfg.clean = True
@@ -111,6 +118,8 @@ def initial_setup(cfg):
 
     change the cwd to the new workdir for this shot
 
+    this is largely equivalent to what rsetups would do, but here for a single shot and not a whole month
+
     :return:
     """
 
@@ -118,24 +127,61 @@ def initial_setup(cfg):
 
     if os.path.exists(workdir):
         if cfg.overwrite:
-            print(f"Overwriting directory {workdir} ... ")
-            shutil.rmtree(workdir)
+            if DEBUG_NO_COPY:
+                print("DEBUG ... skip overwrite directory")
+            else:
+                print(f"Overwriting directory {workdir} ... ")
+                shutil.rmtree(workdir)
         else:
             print(f"Shot directory already exists here! {workdir}")
             return -1
 
     os.makedirs(workdir)
+
+    if LocalScriptRepo is not None:
+        if os.path.exists(LocalScriptRepo): #we want to use it
+            print("Using existing local repo ...")
+            cfg.scriptdir = os.path.join(os.getcwd(), LocalScriptRepo)
+        else:
+            #copy first to local script repo
+            print("Copying to local repo ...")
+            shutil.copytree(os.path.join(ScriptRepo, "science_reductions"),
+                            os.path.join(os.getcwd(),LocalScriptRepo), dirs_exist_ok=True)
+            cfg.scriptdir = os.path.join(os.getcwd(), LocalScriptRepo)
+    else:
+        print("Using main script repo (may be remote) ...")
+        cfg.scriptdir = ScriptRepo
+
     os.chdir(workdir)
     cfg.cwd = workdir #now under the sci<shot> directory
 
-    shutil.copy2(os.path.join(ScriptRepo, "science_reductions", "rsetups"),".")
-    shutil.copy2(os.path.join(ScriptRepo, "science_reductions", "rfixspec"),".")
+    if DEBUG_NO_COPY:
+        print("DEBUG ... skip source code copy")
+        return 0
 
-    shutil.copytree(os.path.join(ScriptRepo,"science_reductions","vdrp"), "vdrp", dirs_exist_ok=True)
+    print("Copying source code ...")
+
+    #shutil.copy2(os.path.join(cfg.scriptdir, "science_reductions", "rsetups"),".") #no, this function is its equivalent
+
+    shutil.copy2(os.path.join(cfg.scriptdir, "rfixspec"), ".")
+    shutil.copytree(os.path.join(cfg.scriptdir, "sciscripts"), ".", dirs_exist_ok=True)
+    shutil.copytree(os.path.join(cfg.scriptdir,"vdrp"), "vdrp", dirs_exist_ok=True)
 
 
 
     return 0
+
+
+def run1s():
+    """
+
+    science_reduction/sciscripts/run1  batch file
+
+
+    :return:
+    """
+
+
 
 
 ########################################################################
